requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0.5
django-allauth==0.51.0
django-countries==7.3.2
django-crispy-forms==1.14.0
django-filter==21.1
gunicorn==20.1.0
idna==3.3
oauthlib==3.2.0
Pillow==9.1.1
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
typing-extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.2.0